requirements.txt
streamlit
pandas
scikit-learn
PyPDF2
pdfplumber
python-docx
matplotlib
